fix: Fix Boss Animation
- 보스 총 쏘는 기능과 애니메이션 부조화 수정해서 맞추기

+++ Assets/01.Scripts/Enemy/EnemyWeapon.cs
@@ -4,7 +4,7 @@ public abstract class EnemyWeapon : MonoBehaviour
 {
     [SerializeField] protected int damage;
     [SerializeField] protected int range;
-    [SerializeField] protected int delay;
+    [SerializeField] protected float delay;
     [SerializeField] protected bool isBetterAttack;
 
     protected float distance;
